CUDA difference kernels now use mask image. Now produces more accurate results with custom cell shapes.
Not yet checking for image bounds so cells exceeding the Photomosaic
edge are not accurate.

Other than the edge cases the results are very different from non-CUDA
version (should be identical) so will need to debug.

CUDA bestFit default value now set to number of library images instead
of 0, to allow for detecting when no best fit found.