GridViewer works with custom cell shapes.
A grid is generated for each size step, but only in the bounds of cells
with high entropy.

Grids are combined from smallest to largest using the larger grid as a
mask.

Edge cells display the outer edge which is leading to some cut-off, need
to instead get inner edge.