GridViewer bug fixed better.
The normal sized cell would pass checks but that didn't mean that the
detail cell would. Added check for detail cell being in bounds and if
not it treats as a low entropy cell.

Problem is now the GridViewer creates a fairly significant different
grid to the Photomosaic. I believe the Photomosaic is suffering from a
similar problem which resulted in missing cells.