Fixed bug with mil file loading.
mil file used to use int for library size but was changed to size_t, but
loading was not updated. Now handles size_t and if mil file is version 2
reads int and casts to size_t.

Updated MIL_VERSION.

Adding lots of images to library now less likely to cause 'not
responding'.