Added complementary colour scheme (#20)
It uses HSV_FULL to help maintain precision, but there is a cost to it. The actually operation (rotating hue) is very cheap, most of the time is taken by converting to and from HSV_FULL. We could instead just use HSV to lessen this at the cost of precision.