Added png image encoding for mil and mcs files (#26, #27)
Incremented mcs to v7, mil to v6.

Added CustomQDataStream to replace the cv::Mat QDataStream operators << >> in ImageUtility, allows us to set a mode on the data stream controlling how the cv::Mat are saved/loaded. RAW mode is the old method (no encoding), ENCODE_PNG is the new method.
It's a bit dodgy, a lot of the operators << >> from the inherited QDataStream don't work so we have to cast back to QDataStream to use them.

The .mcs files have a massive size reduction, and the .mil file has a considerable decrease too.

---------------------------------------------------
Start of #22. Added CUDAImageLibrary.
Added ImageLibrary::addImageInternal(...) that adds an image at the given index in relevant containers. We override it in the CUDA version to also upload images to the GPU.

---------------------------------------------------
Extension of 2ffd5bf, applied filters to the test and benchmark projects.

This really should have been multiple commits, but it got away from me a bit.