We now use the CUDAImageLibrary (#22)
Added a version of CUDAPhotomosaicGenerator::copyMatToDevice(...) for GpuMat.

We are still copying the image data, but now it is device to device rather than host to device.

So far this seems to have had no impact on the run time which is unfortunate but not all that surprising. There are still lots of potential improvments to make, and hopefully when I'm done it will be faster.

-------------------------------------

Removed the OPENCV_W_CUDA define, now CUDA builds always use OpenCV CUDA stuff and non-CUDA builds never do.

Removed unused PhotomosaicGeneratorBase::resizeAndCvtColor() (it was replaced by the preprocess main/library functions).