We now directly use the library image data from the GpuMat in the kernels. (#22)
Added a check that all the library images are continuous (they should always be, but this is safer and gives clear intent).

Previously we had one continuous block of memory for all of the library images, now we have an array of pointers to all the library images. The difference kernels have been changed to handle the array of pointers to library images.

WARNING: During a short performance test I noticed a deviation in the photomosaic result (test case 3). Next step I need to fix the tests to work with these changes and find what caused this change.

------------------------

Removed the CUDAPhotomosaicGenerator::copyMatToDevice(...) for GpuMat as I already stopped using it.