Reverting changes for #22 (some of the changes from f188581, ff0c1ad, b431d9a, 5d29bb1, 2a423e9, dd12a0f, 56d2391, 0dc78a0, 6892b25).
We still have the improvements to the tests, and the difference kernel operates on a single library image. The CPU vs CUDA generator tests are now passing again.

-------------------------------------------------------

Added a second path to CUDAPhotomosaicGenerator::copyMatToDevice(...). If the cv::Mat is continuous then we can copy all the data at once, instead of a row at a time.