Re-working the colour difference kernels #28
Moved the actual core operation to ColourDifference.cuh.
PhotomosaicGenerator.cu(h) takes the images and passes the values to ColourDifference.cuh.

Split the difference kernels in PhotomosaicGenerator.cu(h) into two. We had some code that handled the edge cases (cells that exceed the photomosaic), but most cells didn't need it. So we have one for normal cells and another for the edge cases. The host decides which is needed.

Removed the channels parameter from kernels, we always have 3-channel images so no point to it.
We were multiplying multiple variables by the channels, then in some places dividing it. Now we only multiple the image indices we pass to the ColourDifference kernels.

The target_area (edge case) is now a single block. It should optimise better (or at least have the potential to) as the divide and modulo can be calculated from a single instruction.

ColourDifference::getCUDAFunction(...) now also takes a bool to say if you want the edge case kernel.

tst_ColourDifference only tests the non-edge case kernels, I should add a test for the edge case kernel.

------------------------------
Fixed the canceling in CUDAPhotomosaicGenerator. Now it actually cleans up the CUDA memory.

Added some casts to fix warnings (ImageLibraryEditor, CustomQDataStream.

Added a missing ifdef in TimingLogger.