CUDA math and branch diversion optimisations #28
Re-worked the imageDifference and imageDifferenceEdge kernels in an attempt to minimise branch diversion.

Changed the maths in the colour difference kernels.
Changed all literals to doubles to avoid casting.
Expanded pow() with literal exponents, the pow function is expensive so we can just multiply ourselves (at least for smaller exponents).

Given the complexity of the CIEDE2000Difference kernel really I need to create separate kernels for testing each change to see if it actually helps. Also there will be some branch diversion in here, we should be able to do something about that.

This has given a massive speed-up! The kernels are basically 2x faster.

-------------------------------------------
Added average duration to the TimingLogger summary output.