CUDAPhotomosaicGenerator now works with colour schemes (#20)
We now create device memory for the variants and cell image for each colour scheme variation.
Also renamed some variables; a host vector of device pointers is called h_d%.

Added consistency and comparison test cases for all colour schemes.

Modified the calculateRepeats and findLowestKernel kernels to work on multiple variants arrays (where each one is from a colour scheme variation).

---------------------------------------------
Added flattenKernel to replace the shifting of reduced variants in CUDAPhotomosaicGenerator with cudaMemcpyAsync. (#28)

Added a test for the flattenKernel.

Removed the clearing of the reduction memory, we don't actually need to clear it.

Best fits are now copied at the end of each size step instead of each x loop.