(#24) GridEditor, ColourVisualisation, ImageSquarer, and PhotomosaicViewer now log on open and close.
Added some logging to the PhotomosaicGenerator classes at the key stages (preprocessing, generation start, cancellation, CUDA memory allocation, CUDA memory freeing).

(#25)
Started improving the CUDA error recovery. Currently just in MainWindow::CUDAinit() and ::CUDADeviceChanged().

Added CUDAUtility::CUDAErrMessageBox(). For each call the cudaError codes are put in three categories: SUCCESS, ERR_EXPECTED, and ERR_UNEXPECTED. The expected errors are passed in the parameters.
If it is an error then we display a critical message box containing the CUDA error string, file, and line.
For unexpected errors we also ask the user to report it and include a link to the github issues page.
Might end up changing this...

For the CUDA device combo box we now set the item data to the device index. MainWindow::CUDADeviceChanged() then retrieves the data when changing the device.
Fixes a bug where the wrong device would be set if not all were added to the combo box.
If ::CUDADeviceChanged() has any errors then it will try to set another device. If no devices remain then the CUDA functionality is disabled.