(#25) Improved CUDA error reporting and recovery.
Moved the device memory pointers in CUDAPhotomosaicGenerator::generateBestFits() to member variables. The memory allocation and freeing are now in their own functions ::allocateDeviceMemory() and ::freeDeviceMemory().

gpuErrchk is now only used in the test project for the individual kernel tests, everywhere else we now report the error and try to clean up.

We now pass the selected device to CUDAPhotomosaicGenerator so it reads the correct max block size.