(#25) We now track the created and destroyed CUDA streams. When destroying, we only log the failures and only show a message box after we've attempted to destroy them all. The exceptions are: - cudaErrorMemoryAllocation on cudaMalloc() calls. - cudaErrorInvalidDevice on the cudaGetDeviceProperties() call in CUDAPhotomosaicGenerator::generateBestFits() since we pre-select the device this should be safe?
Removed the expected error codes for most of the CUDA checks as the only way for the errors to occur are errors in the code.

Re-arranged CUDAPhotomosaicGenerator::freeDeviceMemory() to match the same order we allocate in.

-----------------------------------------------------------------
Fixed warning and critical message box initial log types.

Fixed QString arg uses. You can replace multiple arguments at a time except sometimes it decides to do weird implicit casts to match alternate arg methods... so just always do one at a time to be safe.